requirements.txt
gensim>=4.1.2
numpy>=1.22.3
pandas>=1.4.1
progressbar>=2.5
scikit-learn>=1.0.2
termcolor>=1.1.0
torch>=1.11.0
tqdm>=4.63.0
transformers=4.17.0
ufal.udpipe>=1.2.0.3
